+++ test/components/Pagination/Pagination.test.tsx
@@ -10,14 +10,13 @@ describe('Pagination', () => {
     });
 
     it('should render a nav with .pagination and modifiers', () => {
-        const component = shallow(<Pagination isMedium isCentered>My Pagination</Pagination>);
+        const component = shallow(<Pagination isSize='medium'>My Pagination</Pagination>);
         expect(component.hasClass('pagination')).toBe(true);
         expect(component.hasClass('is-medium')).toBe(true);
-        expect(component.hasClass('is-centered')).toBe(true);
     });
 
     it('should render a nav with .pagination and custom classNames', () => {
-        const component = shallow(<Pagination isMedium className='custom'><span>Any Content</span></Pagination>);
+        const component = shallow(<Pagination isSize='medium' className='custom'><span>Any Content</span></Pagination>);
         expect(component.hasClass('pagination')).toBe(true);
         expect(component.hasClass('is-medium')).toBe(true);
         expect(component.hasClass('custom')).toBe(true);